+++ leia-models/src/main/java/com/grookage/leia/models/mux/MessageRequest.java
@@ -20,4 +20,6 @@ public class MessageRequest {
 	private SchemaKey schemaKey;
 	@NotNull
 	private JsonNode message;
+	@NotEmpty
+	private String sourceId;
 }