+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.1.1]
 - `DefaultLeiaMessageValidator`- Bug fix: Fixed validation exception being thrown for optional attributes with null values
+- Added a sourceId field in LeiaMessage
 ## [1.1.0]
 
 - Minor Upgrade. Changed contract in LeiaMessageProduceClient, needs to take in a MessageRequest, instead of